feat: revamp store settings (#239)
